ocr/requirements.txt
opencv-python>=4.5.0
pytesseract>=0.3.8
Pillow>=8.0.0
numpy>=1.20.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
reportlab>=3.6.0
